refactor: 코드 일관성 점검 및 개발 환경 설정 정리
- 미사용 유틸 함수 제거, 상수 접근 제어자 private으로 통일
- 불필요한 주석 제거, 게임 요구사항 맥락 주석 추가
- EOL 설정 통일 (endOfLine: lf, .gitattributes 추가)
- 폴더명 소문자 통일 (Domain→domain, View→view)
- build-dev 스크립트 제거, 빌드 스크립트 단일화
