+++ src/domain/DomainError.js
