+++ src/moveStrategies.js
@@ -1,3 +1,4 @@
+// 미션 요구사항: 0-9 사이의 임의의 숫자를 생성하여, 4 이상일 때 전진
 const MIN_NUMBER = 0;
 const MAX_NUMBER = 9;
 const generateRandomNumber = () =>
